Found out that Utah was missing
So somehow Utah was missing from the poverty CSV.  I updated the CSV, but named it the exact same so no code should need changed in anyone's file that was reading it in or using the dataframe.  You may want to rerun your graphics if you had something that displayed all states, but that just means re-running.